+++ src/MusicalScore/ScoreIO/InstrumentReader.ts
@@ -473,6 +473,9 @@ export class InstrumentReader {
              this.currentMeasure.endsPiece = true;
            }
           }
+          if (xmlNode.element("bar-style") !== undefined) {
+            this.currentMeasure.endingBarStyle = xmlNode.element("bar-style").value;
+          }
         } else if (xmlNode.name === "sound") {
           // (*) MetronomeReader.readTempoInstruction(xmlNode, this.musicSheet, this.currentXmlMeasureIndex);
         } else if (xmlNode.name === "harmony") {
@@ -499,8 +502,16 @@ export class InstrumentReader {
          const reader: ExpressionReader = this.expressionReaders[i];
          if (reader !== undefined) {
            reader.checkForOpenExpressions(this.currentMeasure, currentFraction);
-      }
+          }
         }
+      }
+
+      // if this is the first measure and no BPM info found, we set it to 120
+      // next measures will automatically inherit that value
+      if (!this.musicSheet.HasBPMInfo) {
+        this.currentMeasure.TempoInBPM = 120;
+      } else if (currentMeasure.TempoInBPM === 0) {
+        this.currentMeasure.TempoInBPM = this.previousMeasure.TempoInBPM;
       }
     } catch (e) {
       if (divisionsException) {